refactor: renaming files
- Using `index.js` for root components within directories
- Using `schema.js` for schemas within component directories
