+++ app/components/pages/submission-wizard/steps/Author/schema.js
