+++ app/components/pages/submission-wizard/steps/Editors/index.js
