+++ app/components/pages/submission-wizard/steps/Editors/schema.js
