+++ app/components/pages/submission-wizard/steps/Files/index.js
