+++ app/components/pages/submission-wizard/steps/Files/schema.js
